Transform a spec_df Object into a spec Object or List of spec Objects

Transform a spec_df Object into a spec Object or List of spec Objects

Usage

DF2Spec(spec_df)

Arguments

spec_df

A spec_df object

Value

A spec object or a list of spec objects

See also

Author

Andrew Dolman <andrew.dolman@awi.de>

Examples

library(PaleoSpec)
ts1 <- ts(SimPLS(1000, beta = 1))
sp1 <- SpecMTM(ts1)
ts2 <- ts(rnorm(1000))
sp2 <- SpecMTM(ts2)
sp_lst <- list(sp1 = sp1, sp2 = sp2)
sp_df <- Spec2DF(sp_lst)
sp_df
#> # A tibble: 1,000 × 4
#>    spec_id  freq  spec   dof
#>    <chr>   <dbl> <dbl> <dbl>
#>  1 sp1     0.001 13.6   5.99
#>  2 sp1     0.002 20.2   5.99
#>  3 sp1     0.003 10.2   5.99
#>  4 sp1     0.004 11.5   5.99
#>  5 sp1     0.005  8.41  5.97
#>  6 sp1     0.006  2.68  5.95
#>  7 sp1     0.007  2.55  5.88
#>  8 sp1     0.008  6.24  5.96
#>  9 sp1     0.009  9.54  5.99
#> 10 sp1     0.01   8.86  5.98
#> # ℹ 990 more rows
str(DF2Spec(sp_df))
#> List of 2
#>  $ sp1:List of 4
#>   ..$ spec_id: chr [1:500] "sp1" "sp1" "sp1" "sp1" ...
#>   ..$ freq   : num [1:500] 0.001 0.002 0.003 0.004 0.005 0.006 0.007 0.008 0.009 0.01 ...
#>   ..$ spec   : num [1:500] 13.57 20.19 10.18 11.52 8.41 ...
#>   ..$ dof    : num [1:500] 5.99 5.99 5.99 5.99 5.97 ...
#>   ..- attr(*, "class")= chr [1:2] "spec" "list"
#>  $ sp2:List of 4
#>   ..$ spec_id: chr [1:500] "sp2" "sp2" "sp2" "sp2" ...
#>   ..$ freq   : num [1:500] 0.001 0.002 0.003 0.004 0.005 0.006 0.007 0.008 0.009 0.01 ...
#>   ..$ spec   : num [1:500] 0.69 1.148 1.694 0.862 1.273 ...
#>   ..$ dof    : num [1:500] 5.75 5.84 5.9 5.79 5.86 ...
#>   ..- attr(*, "class")= chr [1:2] "spec" "list"
#>  - attr(*, "class")= chr [1:2] "list" "spec_list"