Transform a spec_df Object into a spec Object or List of spec Objects

Transform a spec_df Object into a spec Object or List of spec Objects

Usage

DF2Spec(spec_df)

Arguments

spec_df

Value

A spec object or a list of spec objects

See also

Author

Andrew Dolman <andrew.dolman@awi.de>

Examples

library(PaleoSpec)
ts1 <- ts(SimPLS(1000, beta = 1))
sp1 <- SpecMTM(ts1)
ts2 <- ts(rnorm(1000))
sp2 <- SpecMTM(ts2)
sp_lst <- list(sp1 = sp1, sp2 = sp2)
sp_df <- Spec2DF(sp_lst)
#> Loading required package: tibble
sp_df
#> # A tibble: 1,000 × 4
#>    spec_id  freq  spec   dof
#>    <chr>   <dbl> <dbl> <dbl>
#>  1 sp1     0.001 22.5   5.99
#>  2 sp1     0.002 34.2   5.99
#>  3 sp1     0.003 14.6   5.99
#>  4 sp1     0.004 15.1   5.99
#>  5 sp1     0.005 12.5   5.98
#>  6 sp1     0.006 15.9   5.98
#>  7 sp1     0.007 11.0   5.98
#>  8 sp1     0.008  7.15  5.98
#>  9 sp1     0.009  4.79  5.95
#> 10 sp1     0.01  10.1   5.98
#> # … with 990 more rows
DF2Spec(sp_df)
#> $sp1
#> $spec_id
#>   [1] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [13] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [25] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [37] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [49] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [61] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [73] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [85] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [97] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [109] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [121] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [133] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [145] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [157] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [169] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [181] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [193] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [205] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [217] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [229] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [241] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [253] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [265] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [277] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [289] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [301] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [313] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [325] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [337] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [349] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [361] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [373] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [385] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [397] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [409] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [421] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [433] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [445] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [457] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [469] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [481] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [493] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> 
#> $freq
#>   [1] 0.001 0.002 0.003 0.004 0.005 0.006 0.007 0.008 0.009 0.010 0.011 0.012
#>  [13] 0.013 0.014 0.015 0.016 0.017 0.018 0.019 0.020 0.021 0.022 0.023 0.024
#>  [25] 0.025 0.026 0.027 0.028 0.029 0.030 0.031 0.032 0.033 0.034 0.035 0.036
#>  [37] 0.037 0.038 0.039 0.040 0.041 0.042 0.043 0.044 0.045 0.046 0.047 0.048
#>  [49] 0.049 0.050 0.051 0.052 0.053 0.054 0.055 0.056 0.057 0.058 0.059 0.060
#>  [61] 0.061 0.062 0.063 0.064 0.065 0.066 0.067 0.068 0.069 0.070 0.071 0.072
#>  [73] 0.073 0.074 0.075 0.076 0.077 0.078 0.079 0.080 0.081 0.082 0.083 0.084
#>  [85] 0.085 0.086 0.087 0.088 0.089 0.090 0.091 0.092 0.093 0.094 0.095 0.096
#>  [97] 0.097 0.098 0.099 0.100 0.101 0.102 0.103 0.104 0.105 0.106 0.107 0.108
#> [109] 0.109 0.110 0.111 0.112 0.113 0.114 0.115 0.116 0.117 0.118 0.119 0.120
#> [121] 0.121 0.122 0.123 0.124 0.125 0.126 0.127 0.128 0.129 0.130 0.131 0.132
#> [133] 0.133 0.134 0.135 0.136 0.137 0.138 0.139 0.140 0.141 0.142 0.143 0.144
#> [145] 0.145 0.146 0.147 0.148 0.149 0.150 0.151 0.152 0.153 0.154 0.155 0.156
#> [157] 0.157 0.158 0.159 0.160 0.161 0.162 0.163 0.164 0.165 0.166 0.167 0.168
#> [169] 0.169 0.170 0.171 0.172 0.173 0.174 0.175 0.176 0.177 0.178 0.179 0.180
#> [181] 0.181 0.182 0.183 0.184 0.185 0.186 0.187 0.188 0.189 0.190 0.191 0.192
#> [193] 0.193 0.194 0.195 0.196 0.197 0.198 0.199 0.200 0.201 0.202 0.203 0.204
#> [205] 0.205 0.206 0.207 0.208 0.209 0.210 0.211 0.212 0.213 0.214 0.215 0.216
#> [217] 0.217 0.218 0.219 0.220 0.221 0.222 0.223 0.224 0.225 0.226 0.227 0.228
#> [229] 0.229 0.230 0.231 0.232 0.233 0.234 0.235 0.236 0.237 0.238 0.239 0.240
#> [241] 0.241 0.242 0.243 0.244 0.245 0.246 0.247 0.248 0.249 0.250 0.251 0.252
#> [253] 0.253 0.254 0.255 0.256 0.257 0.258 0.259 0.260 0.261 0.262 0.263 0.264
#> [265] 0.265 0.266 0.267 0.268 0.269 0.270 0.271 0.272 0.273 0.274 0.275 0.276
#> [277] 0.277 0.278 0.279 0.280 0.281 0.282 0.283 0.284 0.285 0.286 0.287 0.288
#> [289] 0.289 0.290 0.291 0.292 0.293 0.294 0.295 0.296 0.297 0.298 0.299 0.300
#> [301] 0.301 0.302 0.303 0.304 0.305 0.306 0.307 0.308 0.309 0.310 0.311 0.312
#> [313] 0.313 0.314 0.315 0.316 0.317 0.318 0.319 0.320 0.321 0.322 0.323 0.324
#> [325] 0.325 0.326 0.327 0.328 0.329 0.330 0.331 0.332 0.333 0.334 0.335 0.336
#> [337] 0.337 0.338 0.339 0.340 0.341 0.342 0.343 0.344 0.345 0.346 0.347 0.348
#> [349] 0.349 0.350 0.351 0.352 0.353 0.354 0.355 0.356 0.357 0.358 0.359 0.360
#> [361] 0.361 0.362 0.363 0.364 0.365 0.366 0.367 0.368 0.369 0.370 0.371 0.372
#> [373] 0.373 0.374 0.375 0.376 0.377 0.378 0.379 0.380 0.381 0.382 0.383 0.384
#> [385] 0.385 0.386 0.387 0.388 0.389 0.390 0.391 0.392 0.393 0.394 0.395 0.396
#> [397] 0.397 0.398 0.399 0.400 0.401 0.402 0.403 0.404 0.405 0.406 0.407 0.408
#> [409] 0.409 0.410 0.411 0.412 0.413 0.414 0.415 0.416 0.417 0.418 0.419 0.420
#> [421] 0.421 0.422 0.423 0.424 0.425 0.426 0.427 0.428 0.429 0.430 0.431 0.432
#> [433] 0.433 0.434 0.435 0.436 0.437 0.438 0.439 0.440 0.441 0.442 0.443 0.444
#> [445] 0.445 0.446 0.447 0.448 0.449 0.450 0.451 0.452 0.453 0.454 0.455 0.456
#> [457] 0.457 0.458 0.459 0.460 0.461 0.462 0.463 0.464 0.465 0.466 0.467 0.468
#> [469] 0.469 0.470 0.471 0.472 0.473 0.474 0.475 0.476 0.477 0.478 0.479 0.480
#> [481] 0.481 0.482 0.483 0.484 0.485 0.486 0.487 0.488 0.489 0.490 0.491 0.492
#> [493] 0.493 0.494 0.495 0.496 0.497 0.498 0.499 0.500
#> 
#> $spec
#>   [1] 22.51719501 34.15118771 14.64975451 15.08650847 12.49117889 15.93386588
#>   [7] 10.96062496  7.15496126  4.79463310 10.14232674 10.12687508  7.48599005
#>  [13]  4.81429266  5.68325006  3.89894193  2.97181496  5.30598609  5.21488604
#>  [19]  5.53051500  4.49534337  2.46187463  2.95775060  3.19450281  3.85626854
#>  [25]  3.62794903  4.49799411  7.18655418  5.20035974  4.46763198  3.33768168
#>  [31]  1.10243936  3.17216830  2.36124596  1.93532730  1.77287204  0.70142888
#>  [37]  1.15674754  1.14479576  0.75534830  0.45600492  1.61087292  1.99745363
#>  [43]  1.71154518  0.55833448  0.81264864  1.41426581  0.87696207  0.89095548
#>  [49]  0.77416998  0.52456495  0.86240591  1.14606987  1.59738272  1.10206540
#>  [55]  0.58622402  0.74896614  0.52399793  1.02449112  0.56938678  0.72847241
#>  [61]  0.68556284  1.40112687  1.55858946  1.99697195  1.02438260  1.41331661
#>  [67]  0.86238520  0.91232440  0.91584123  0.71329222  0.47096442  0.46662537
#>  [73]  1.05829645  2.16020556  2.71138562  2.57513317  0.69341532  0.78499070
#>  [79]  0.30241949  0.54080909  0.51184626  0.80625836  0.92040274  0.91053036
#>  [85]  0.90361592  0.71658023  1.00264503  0.34592689  0.32138264  0.27483465
#>  [91]  0.23667326  0.43382394  0.48903319  0.43204597  1.51716189  1.27299284
#>  [97]  1.49695029  0.48497034  0.43954774  0.57381073  0.84923644  1.24765481
#> [103]  0.99294291  0.31278022  0.15199947  0.34427980  0.43982855  0.35662633
#> [109]  0.34842256  0.21259987  0.20670875  0.20728369  0.09210225  0.13870205
#> [115]  0.32652114  0.53343256  0.45448162  0.63153534  0.62052874  0.60504683
#> [121]  0.47979552  0.94123964  0.88373859  0.99147699  0.77099979  0.88703630
#> [127]  0.55736207  0.88107072  0.62550956  0.74187741  0.57283003  0.52697052
#> [133]  1.00594997  0.50777871  0.52516092  0.10895752  0.27363377  0.51489716
#> [139]  0.54724171  0.33649739  0.23789694  0.30881404  0.45665612  0.64335817
#> [145]  0.61593658  0.46354171  0.56962781  0.48999238  0.59275262  0.31603839
#> [151]  0.33654302  0.27060577  0.72035606  0.87757813  0.84482242  0.37068797
#> [157]  0.51589218  0.71005005  0.57883700  0.46508209  0.36205457  0.37655289
#> [163]  0.27723176  0.24473292  0.15418920  0.15220485  0.31124505  0.30467009
#> [169]  0.23319401  0.18040291  0.22528301  0.51233807  0.48455683  1.18250477
#> [175]  0.80811703  0.81896238  0.21143674  0.18442178  0.31648362  0.37516061
#> [181]  0.36094556  0.34967816  0.40278049  0.57932048  0.35283295  0.22859585
#> [187]  0.56524517  0.61213903  0.50314606  0.34750084  0.44130269  0.40767003
#> [193]  0.46280688  0.30563094  0.27645807  0.09653856  0.14658356  0.20371666
#> [199]  0.38059020  0.35500232  0.34597435  0.35799841  0.44578368  0.63814587
#> [205]  0.51619821  0.67121723  0.69008878  0.37217139  0.34358675  0.46204709
#> [211]  0.48406504  0.67394833  0.55852374  0.39833685  0.54867084  0.27585845
#> [217]  0.36168969  0.44643799  0.37522312  0.23806748  0.13745004  0.08650142
#> [223]  0.07743429  0.03131602  0.26762049  0.60119232  0.65260695  0.58571493
#> [229]  0.26146870  0.27192033  0.15613487  0.12534720  0.07281234  0.04691487
#> [235]  0.15119596  0.26036640  0.42316154  0.74643902  0.75088787  0.53727122
#> [241]  0.21761981  0.06518003  0.05360470  0.03117038  0.18078015  0.25539601
#> [247]  0.27423374  0.69958552  0.74442305  0.62178208  0.42271424  0.17715148
#> [253]  0.34337314  0.60939869  0.82236784  1.01961330  1.04429782  1.15321009
#> [259]  0.49058305  0.30610319  0.09695681  0.09327390  0.07477594  0.04571455
#> [265]  0.06176562  0.43146736  0.47154369  0.43309015  0.25173359  0.08720429
#> [271]  0.32676114  0.37770703  0.42529446  0.29886535  0.11968757  0.09494493
#> [277]  0.23273550  0.24888985  0.23901425  0.09718191  0.09380612  0.39138416
#> [283]  0.58822079  0.60452921  0.14709387  0.15636208  0.13598012  0.20868831
#> [289]  0.18446881  0.29702694  0.21800559  0.16614093  0.07231376  0.10216504
#> [295]  0.22103824  0.25893804  0.31990696  0.25538289  0.20144118  0.23868755
#> [301]  0.20430453  0.32812652  0.32837425  0.34980023  0.15338646  0.18722417
#> [307]  0.13619576  0.19544194  0.23349005  0.29574364  0.38870897  0.32476162
#> [313]  0.44231501  0.40706319  0.29905802  0.23325569  0.12414802  0.07940845
#> [319]  0.07665693  0.20774684  0.31921068  0.35647369  0.42671866  0.31341958
#> [325]  0.28731773  0.22952589  0.08617502  0.10100904  0.24384402  0.24278771
#> [331]  0.25439603  0.11980705  0.07583133  0.09005169  0.14725301  0.18370200
#> [337]  0.41501721  0.44020948  0.37817628  0.31868057  0.24006980  0.20902090
#> [343]  0.22524266  0.09828078  0.04259908  0.04244040  0.06021366  0.08583211
#> [349]  0.22471952  0.21920980  0.21791564  0.23610187  0.30338636  0.26497408
#> [355]  0.10836620  0.07750039  0.13830872  0.19879960  0.20644284  0.20182749
#> [361]  0.14944341  0.22164006  0.25279976  0.25055158  0.14318468  0.13095498
#> [367]  0.17526864  0.26076760  0.30041774  0.21936857  0.16701707  0.19476218
#> [373]  0.18614112  0.10529649  0.05974242  0.06977262  0.09514246  0.06921002
#> [379]  0.16515394  0.22073971  0.32788915  0.48332298  0.38960801  0.18517033
#> [385]  0.04335958  0.13640784  0.11878696  0.14003684  0.16423756  0.08807390
#> [391]  0.13688036  0.10488146  0.11313842  0.21364157  0.29783488  0.27193631
#> [397]  0.13200804  0.10314997  0.31937217  0.40246457  0.41991810  0.26358185
#> [403]  0.28025266  0.26602126  0.18342205  0.33559097  0.31072547  0.36662833
#> [409]  0.34086963  0.42180823  0.25571481  0.27477959  0.21675670  0.15303987
#> [415]  0.18008616  0.35823984  0.31047053  0.32807166  0.22285685  0.11820557
#> [421]  0.02627724  0.08213721  0.18492858  0.33956942  0.29270096  0.35315056
#> [427]  0.21875376  0.22250147  0.15586183  0.07630620  0.18650163  0.31302252
#> [433]  0.28534123  0.24040880  0.16259921  0.31069327  0.35452854  0.31619363
#> [439]  0.22173570  0.10294192  0.10468547  0.04823767  0.05572768  0.03369261
#> [445]  0.03991801  0.03144791  0.04121675  0.04473230  0.09729397  0.17382429
#> [451]  0.19788073  0.14779529  0.07456556  0.11311198  0.17417394  0.28585498
#> [457]  0.27258100  0.24133291  0.07137726  0.05011674  0.20407982  0.30300765
#> [463]  0.33756869  0.29094492  0.12819358  0.15288980  0.12048349  0.10555634
#> [469]  0.08577486  0.08525844  0.08370117  0.10137999  0.09120442  0.07421270
#> [475]  0.05457791  0.06645488  0.18943584  0.19296061  0.23445765  0.11699969
#> [481]  0.12207525  0.08985345  0.08658657  0.12157023  0.13808062  0.24072990
#> [487]  0.19446934  0.18688959  0.04583439  0.02957488  0.03434030  0.06731736
#> [493]  0.07802733  0.08939760  0.09925402  0.08489220  0.10772017  0.10653904
#> [499]  0.12547844  0.09642453
#> 
#> $dof
#>   [1] 5.985449 5.985088 5.989662 5.985614 5.984893 5.983441 5.981708 5.975530
#>   [9] 5.953432 5.976300 5.987125 5.982906 5.972399 5.927591 5.952118 5.953591
#>  [17] 5.971101 5.975554 5.974313 5.954065 5.916164 5.957900 5.961141 5.967873
#>  [25] 5.964915 5.947051 5.967583 5.976257 5.969527 5.944664 5.842414 5.916665
#>  [33] 5.939239 5.932045 5.899959 5.758077 5.849252 5.847808 5.775101 5.645858
#>  [41] 5.890159 5.912007 5.914584 5.703958 5.790179 5.875544 5.803537 5.806834
#>  [49] 5.778739 5.686221 5.800836 5.880571 5.888962 5.842239 5.716929 5.773583
#>  [57] 5.683259 5.830104 5.709383 5.767705 5.752719 5.898974 5.893304 5.931936
#>  [65] 5.830296 5.875450 5.800676 5.810943 5.811562 5.763336 5.655801 5.652463
#>  [73] 5.835786 5.943482 5.933881 5.935334 5.756848 5.782914 5.495438 5.693909
#>  [81] 5.679420 5.788255 5.812440 5.810625 5.809052 5.764238 5.827544 5.550243
#>  [89] 5.520045 5.451540 5.384707 5.628164 5.666688 5.628716 5.883731 5.887649
#>  [97] 5.882194 5.663333 5.633447 5.711403 5.798354 5.901108 5.825019 5.510277
#> [105] 5.145783 5.547403 5.633714 5.561246 5.547685 5.330621 5.316427 5.319641
#> [113] 4.819454 5.090048 5.527479 5.690893 5.641874 5.734929 5.730995 5.723969
#> [121] 5.658355 5.816980 5.805986 5.824678 5.778862 5.805684 5.700083 5.803946
#> [129] 5.732159 5.771722 5.710717 5.686747 5.827921 5.677973 5.687443 4.933405
#> [137] 5.454305 5.680449 5.698027 5.540180 5.389921 5.504970 5.646306 5.739807
#> [145] 5.728472 5.648529 5.707847 5.664634 5.718498 5.514282 5.539387 5.448949
#> [153] 5.764168 5.804715 5.797622 5.576104 5.681262 5.760010 5.711450 5.651965
#> [161] 5.567394 5.580847 5.460136 5.395433 5.148755 5.147726 5.508465 5.497811
#> [169] 5.379249 5.245987 5.361996 5.679814 5.661861 5.852520 5.788883 5.791719
#> [177] 5.319717 5.254272 5.515923 5.580087 5.564187 5.553687 5.603514 5.713814
#> [185] 5.557642 5.370100 5.706653 5.727758 5.674119 5.551348 5.635519 5.609279
#> [193] 5.649732 5.496552 5.454785 4.836659 5.121241 5.310884 5.583236 5.554863
#> [201] 5.549452 5.563597 5.637693 5.735270 5.680107 5.749606 5.753226 5.577550
#> [209] 5.547865 5.648263 5.663915 5.747897 5.703520 5.595681 5.699227 5.457214
#> [217] 5.567159 5.637891 5.580530 5.389750 5.083559 4.773761 4.695649 4.053864
#> [225] 5.443068 5.722652 5.743086 5.716604 5.431377 5.450872 5.162793 5.024143
#> [233] 4.653147 4.322517 5.141904 5.422343 5.621894 5.772947 5.774152 5.693527
#> [241] 5.344356 4.567398 4.430961 4.035008 5.245357 5.422491 5.455098 5.758673
#> [249] 5.770842 5.730285 5.621094 5.235049 5.544581 5.724892 5.792322 5.830019
#> [257] 5.833633 5.848978 5.667936 5.501821 4.852975 4.823532 4.664225 4.317072
#> [265] 4.532159 5.628264 5.654858 5.625511 5.409976 4.776335 5.528002 5.582490
#> [273] 5.623562 5.489638 4.995582 4.836948 5.378703 5.411434 5.388570 4.847147
#> [281] 4.830792 5.594062 5.714885 5.724523 5.126193 5.149249 5.073947 5.322396
#> [289] 5.258890 5.485865 5.342770 5.196579 4.648565 4.889641 5.351793 5.424239
#> [297] 5.519902 5.418845 5.305796 5.390559 5.311509 5.526512 5.530044 5.553827
#> [305] 5.151895 5.265153 5.078769 5.287956 5.375862 5.486011 5.593352 5.526312
#> [313] 5.634121 5.608836 5.489394 5.380027 5.019996 4.713882 4.690709 5.318087
#> [321] 5.519349 5.562210 5.624468 5.508872 5.471362 5.362833 4.772714 4.881793
#> [329] 5.400246 5.397726 5.417468 4.995175 4.681949 4.801573 5.127101 5.256396
#> [337] 5.612869 5.631488 5.583693 5.517609 5.393606 5.323072 5.362288 4.865398
#> [345] 4.259843 4.268018 4.516226 4.763988 5.360472 5.346259 5.341859 5.384624
#> [353] 5.495503 5.435669 4.930546 4.697768 5.083844 5.297219 5.317675 5.304840
#> [361] 5.136992 5.354857 5.417769 5.406066 5.106713 5.053429 5.224420 5.427914
#> [369] 5.491871 5.342557 5.200122 5.287433 5.263663 4.908563 4.507250 4.619133
#> [377] 4.831114 4.616460 5.193569 5.350591 5.526359 5.663595 5.592393 5.257470
#> [385] 4.278360 5.079701 4.991405 5.095840 5.190778 4.786760 5.079671 4.895215
#> [393] 4.958823 5.335866 5.486126 5.450396 5.058249 4.896316 5.519096 5.601549
#> [401] 5.617591 5.436215 5.464726 5.432136 5.251661 5.537934 5.504831 5.571096
#> [409] 5.543167 5.618208 5.423955 5.455093 5.342763 5.145801 5.244669 5.562936
#> [417] 5.502454 5.527339 5.357247 4.986803 3.928601 4.739359 5.253658 5.540355
#> [425] 5.483933 5.556945 5.347391 5.352099 5.158896 4.683828 5.260290 5.510715
#> [433] 5.472653 5.393910 5.184506 5.508369 5.555653 5.510145 5.354258 4.881016
#> [441] 4.907366 4.354185 4.461980 4.108515 4.226146 4.062338 4.246752 4.304426
#> [449] 4.852143 5.224625 5.295387 5.128273 4.670149 4.958009 5.213733 5.468744
#> [457] 5.451486 5.392614 4.638448 4.383134 5.311984 5.497608 5.540906 5.477817
#> [465] 5.039675 5.149486 4.997027 4.912724 4.765252 4.759384 4.748567 4.884368
#> [473] 4.809854 4.666502 4.444315 4.587180 5.271865 5.282355 5.378388 4.981286
#> [481] 5.007152 4.803024 4.772217 5.005677 5.087782 5.394487 5.287664 5.264913
#> [489] 4.311292 4.021496 4.122136 4.596614 4.702024 4.795900 4.869378 4.759985
#> [497] 4.923734 4.918001 5.022896 4.850607
#> 
#> attr(,".internal.selfref")
#> <pointer: 0x5636ce914340>
#> attr(,"class")
#> [1] "spec" "list"
#> 
#> $sp2
#> $spec_id
#>   [1] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [13] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [25] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [37] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [49] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [61] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [73] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [85] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [97] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [109] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [121] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [133] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [145] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [157] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [169] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [181] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [193] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [205] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [217] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [229] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [241] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [253] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [265] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [277] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [289] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [301] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [313] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [325] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [337] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [349] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [361] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [373] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [385] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [397] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [409] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [421] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [433] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [445] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [457] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [469] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [481] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [493] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> 
#> $freq
#>   [1] 0.001 0.002 0.003 0.004 0.005 0.006 0.007 0.008 0.009 0.010 0.011 0.012
#>  [13] 0.013 0.014 0.015 0.016 0.017 0.018 0.019 0.020 0.021 0.022 0.023 0.024
#>  [25] 0.025 0.026 0.027 0.028 0.029 0.030 0.031 0.032 0.033 0.034 0.035 0.036
#>  [37] 0.037 0.038 0.039 0.040 0.041 0.042 0.043 0.044 0.045 0.046 0.047 0.048
#>  [49] 0.049 0.050 0.051 0.052 0.053 0.054 0.055 0.056 0.057 0.058 0.059 0.060
#>  [61] 0.061 0.062 0.063 0.064 0.065 0.066 0.067 0.068 0.069 0.070 0.071 0.072
#>  [73] 0.073 0.074 0.075 0.076 0.077 0.078 0.079 0.080 0.081 0.082 0.083 0.084
#>  [85] 0.085 0.086 0.087 0.088 0.089 0.090 0.091 0.092 0.093 0.094 0.095 0.096
#>  [97] 0.097 0.098 0.099 0.100 0.101 0.102 0.103 0.104 0.105 0.106 0.107 0.108
#> [109] 0.109 0.110 0.111 0.112 0.113 0.114 0.115 0.116 0.117 0.118 0.119 0.120
#> [121] 0.121 0.122 0.123 0.124 0.125 0.126 0.127 0.128 0.129 0.130 0.131 0.132
#> [133] 0.133 0.134 0.135 0.136 0.137 0.138 0.139 0.140 0.141 0.142 0.143 0.144
#> [145] 0.145 0.146 0.147 0.148 0.149 0.150 0.151 0.152 0.153 0.154 0.155 0.156
#> [157] 0.157 0.158 0.159 0.160 0.161 0.162 0.163 0.164 0.165 0.166 0.167 0.168
#> [169] 0.169 0.170 0.171 0.172 0.173 0.174 0.175 0.176 0.177 0.178 0.179 0.180
#> [181] 0.181 0.182 0.183 0.184 0.185 0.186 0.187 0.188 0.189 0.190 0.191 0.192
#> [193] 0.193 0.194 0.195 0.196 0.197 0.198 0.199 0.200 0.201 0.202 0.203 0.204
#> [205] 0.205 0.206 0.207 0.208 0.209 0.210 0.211 0.212 0.213 0.214 0.215 0.216
#> [217] 0.217 0.218 0.219 0.220 0.221 0.222 0.223 0.224 0.225 0.226 0.227 0.228
#> [229] 0.229 0.230 0.231 0.232 0.233 0.234 0.235 0.236 0.237 0.238 0.239 0.240
#> [241] 0.241 0.242 0.243 0.244 0.245 0.246 0.247 0.248 0.249 0.250 0.251 0.252
#> [253] 0.253 0.254 0.255 0.256 0.257 0.258 0.259 0.260 0.261 0.262 0.263 0.264
#> [265] 0.265 0.266 0.267 0.268 0.269 0.270 0.271 0.272 0.273 0.274 0.275 0.276
#> [277] 0.277 0.278 0.279 0.280 0.281 0.282 0.283 0.284 0.285 0.286 0.287 0.288
#> [289] 0.289 0.290 0.291 0.292 0.293 0.294 0.295 0.296 0.297 0.298 0.299 0.300
#> [301] 0.301 0.302 0.303 0.304 0.305 0.306 0.307 0.308 0.309 0.310 0.311 0.312
#> [313] 0.313 0.314 0.315 0.316 0.317 0.318 0.319 0.320 0.321 0.322 0.323 0.324
#> [325] 0.325 0.326 0.327 0.328 0.329 0.330 0.331 0.332 0.333 0.334 0.335 0.336
#> [337] 0.337 0.338 0.339 0.340 0.341 0.342 0.343 0.344 0.345 0.346 0.347 0.348
#> [349] 0.349 0.350 0.351 0.352 0.353 0.354 0.355 0.356 0.357 0.358 0.359 0.360
#> [361] 0.361 0.362 0.363 0.364 0.365 0.366 0.367 0.368 0.369 0.370 0.371 0.372
#> [373] 0.373 0.374 0.375 0.376 0.377 0.378 0.379 0.380 0.381 0.382 0.383 0.384
#> [385] 0.385 0.386 0.387 0.388 0.389 0.390 0.391 0.392 0.393 0.394 0.395 0.396
#> [397] 0.397 0.398 0.399 0.400 0.401 0.402 0.403 0.404 0.405 0.406 0.407 0.408
#> [409] 0.409 0.410 0.411 0.412 0.413 0.414 0.415 0.416 0.417 0.418 0.419 0.420
#> [421] 0.421 0.422 0.423 0.424 0.425 0.426 0.427 0.428 0.429 0.430 0.431 0.432
#> [433] 0.433 0.434 0.435 0.436 0.437 0.438 0.439 0.440 0.441 0.442 0.443 0.444
#> [445] 0.445 0.446 0.447 0.448 0.449 0.450 0.451 0.452 0.453 0.454 0.455 0.456
#> [457] 0.457 0.458 0.459 0.460 0.461 0.462 0.463 0.464 0.465 0.466 0.467 0.468
#> [469] 0.469 0.470 0.471 0.472 0.473 0.474 0.475 0.476 0.477 0.478 0.479 0.480
#> [481] 0.481 0.482 0.483 0.484 0.485 0.486 0.487 0.488 0.489 0.490 0.491 0.492
#> [493] 0.493 0.494 0.495 0.496 0.497 0.498 0.499 0.500
#> 
#> $spec
#>   [1] 0.25488507 0.81521629 0.89531555 0.64709585 0.96750937 1.14119732
#>   [7] 2.14646446 1.21544938 2.36534659 3.38371357 2.78591817 2.12030037
#>  [13] 1.73404262 1.54235085 2.19190938 2.44097279 2.23887743 1.59941467
#>  [19] 0.62762187 0.43505970 1.17212141 1.32588174 1.82867587 2.52727924
#>  [25] 1.54666767 1.52726031 1.19052072 0.72891423 0.62862010 0.24984009
#>  [31] 0.21270697 0.09033017 0.34180680 0.46038550 1.60282487 1.13610432
#>  [37] 0.79767405 0.20737484 0.45790195 0.84188465 1.08046275 1.03554788
#>  [43] 1.18840391 0.89936823 0.57163311 0.50227108 0.38917837 0.47005463
#>  [49] 0.24591475 0.15871341 0.13271993 0.17341901 1.21343142 1.57705578
#>  [55] 1.37265556 1.41577594 1.32681717 0.91750913 0.87206056 1.02203543
#>  [61] 1.15639711 1.15759255 0.98396812 1.23270484 1.13858299 0.81391105
#>  [67] 1.17725937 0.79057190 0.54674258 0.25969028 0.27083665 0.47674605
#>  [73] 1.16278541 1.00289229 1.96098672 1.68228108 1.99196903 1.20524789
#>  [79] 0.72777414 0.56151241 0.90288302 1.15027291 1.01836961 0.60699345
#>  [85] 1.87659973 1.74587898 1.46264256 0.53646509 0.30108417 0.31470206
#>  [91] 0.84610625 1.11566123 0.89531958 0.35052160 0.57837191 0.44230897
#>  [97] 0.52338224 0.40666310 1.04179996 0.64411364 0.84595616 1.38666590
#> [103] 1.47226313 1.64404190 0.28935180 0.15602467 0.98852476 1.16447540
#> [109] 1.15391755 1.22878965 0.95389668 0.73781989 0.51788354 0.08847954
#> [115] 0.76837991 0.86418860 1.01295772 0.89755029 1.10853159 0.78291472
#> [121] 0.89232005 0.70655234 2.04107395 2.30414323 2.16762126 1.13825957
#> [127] 0.26142755 0.12156315 0.19311049 0.33092107 0.60719284 0.57230687
#> [133] 0.83012016 0.82875012 0.79197327 0.75630597 0.43147455 1.17460454
#> [139] 0.96709726 0.74412314 0.59867614 0.79029305 0.86923337 1.27106311
#> [145] 1.41469540 1.77880959 1.07075508 1.14688729 1.19165824 1.44668853
#> [151] 2.29950548 2.51818667 2.10186018 2.16073635 1.47736005 1.42957016
#> [157] 2.01891989 2.05161341 2.46616022 2.23492088 2.18828564 2.77691591
#> [163] 1.38898087 1.48471787 1.29959297 0.08881336 0.13339152 0.65290933
#> [169] 1.26075865 1.21677896 1.29844636 0.49332149 0.48919358 0.96781303
#> [175] 0.85777043 0.96900081 0.65491282 0.64226217 0.51785443 0.44127362
#> [181] 0.36288898 0.61115813 0.86811449 0.89758625 0.57482051 1.12258814
#> [187] 0.72223883 0.69039554 0.98087867 0.99006228 1.39530127 0.97880821
#> [193] 0.73474757 0.09637927 0.44276167 1.32844159 1.19123379 1.23441326
#> [199] 0.51211747 1.13963089 0.83860475 1.11217386 0.96154426 0.63264086
#> [205] 0.50145465 0.33446203 1.28982663 0.92482127 0.85954969 0.41268897
#> [211] 0.12751230 0.26558973 0.87462329 0.88202724 0.79035003 0.34553349
#> [217] 0.24631315 0.53612315 0.52523022 0.43099178 0.83575590 1.33549842
#> [223] 1.56894673 2.21604324 1.82378147 2.09987178 1.79132315 1.41549154
#> [229] 0.91828382 0.85534518 1.07027294 1.27449328 1.11117339 0.92774493
#> [235] 0.56004572 0.32428383 1.02830889 2.23320739 3.39718591 3.37852350
#> [241] 1.72483054 1.52115870 0.66228646 0.57149332 0.78947190 0.45730608
#> [247] 0.79446460 1.01187830 1.17315170 0.79472148 0.80555902 0.82348521
#> [253] 1.58043857 1.73504873 1.45072834 1.05920246 1.32980735 1.33332577
#> [259] 1.74143348 1.17964683 0.84387055 0.89272891 0.59407895 0.78997030
#> [265] 1.33732160 1.05166940 0.97612050 1.22307072 1.17590991 1.38645187
#> [271] 0.70831227 1.08842310 0.97811159 1.07934819 0.59491918 1.14309627
#> [277] 0.66471305 0.99786479 1.29622061 1.20062246 1.12525920 0.64471256
#> [283] 1.07290926 0.70659457 1.09501172 1.11860699 1.66864834 1.06322184
#> [289] 1.00743440 1.10286714 1.51046234 1.58093379 1.51756742 0.45699355
#> [295] 0.79559923 0.63756439 0.67474329 0.48583366 0.89248228 1.34679737
#> [301] 1.11622551 1.78608822 2.25797856 1.80108624 1.41098571 0.31494947
#> [307] 0.11130049 0.33487119 0.91976450 1.43250740 1.68723954 1.80102111
#> [313] 0.78388330 0.22953455 0.55073519 1.90640904 2.63217754 3.34752371
#> [319] 1.46148170 1.32271358 1.23060923 0.71302086 0.63268562 0.59336972
#> [325] 0.51814087 0.31453782 0.29193652 0.23348193 0.23815780 0.30502405
#> [331] 0.65836343 1.04070118 1.46877641 1.34208710 1.29236509 0.93889873
#> [337] 0.61580128 0.85501870 1.16927324 1.19905669 1.37045465 0.78638384
#> [343] 0.95914678 0.92827697 0.72242896 0.62740003 0.62753969 1.00125755
#> [349] 1.32358396 1.30547328 0.96294170 0.70542997 0.25697058 0.36323913
#> [355] 1.35575105 2.13235357 1.87509510 2.09335947 0.97409869 0.56565982
#> [361] 0.61963498 1.52766980 1.72279299 1.94418189 1.52033999 1.02435082
#> [367] 1.07085399 0.56977554 0.80896970 0.73132093 0.73077378 0.31117148
#> [373] 0.25998728 1.24473481 1.20735615 1.52882050 1.19739903 0.41230347
#> [379] 0.22746983 0.28720020 0.42949221 0.69032186 0.69484581 0.98155480
#> [385] 1.42974416 0.66694239 1.01312311 1.02228869 0.95431233 0.94636293
#> [391] 1.67372283 2.03824535 2.07488255 0.43367722 0.36179214 0.34574051
#> [397] 0.91727595 1.57998349 1.57793148 1.83760576 0.76342864 0.47996954
#> [403] 0.60924438 1.17726253 2.08223379 0.99557451 0.66298730 0.59458302
#> [409] 0.40332592 0.39611946 0.43680367 0.43963515 0.72202692 1.06768719
#> [415] 1.27421084 0.75591421 0.57636344 0.36840804 0.59362311 0.74304266
#> [421] 1.24364413 1.40238486 1.44597017 1.37916641 1.15889818 1.40080353
#> [427] 1.42259742 0.50378018 0.24916623 0.54661861 1.01525762 1.17310439
#> [433] 1.30380896 1.27514897 0.53522572 0.33000084 0.30105245 0.32362404
#> [439] 0.42440090 0.73146346 1.22621309 1.36770244 0.82284987 0.60884734
#> [445] 0.70345676 0.60725020 1.10329670 0.59737612 0.49416302 0.25189872
#> [451] 0.84289417 0.88963574 0.56616476 0.48764126 0.38345110 0.43748448
#> [457] 0.95450884 0.92370076 0.93835335 0.39824467 0.23685895 0.32033948
#> [463] 0.34605355 0.64551083 1.05082453 1.22814520 0.91136163 1.55335733
#> [469] 1.09897876 1.04445484 0.45728188 1.03430574 1.31726309 2.10395314
#> [475] 1.76221845 1.20607509 0.91890608 0.34768187 0.32315889 0.53666479
#> [481] 0.82892140 1.16010728 0.92300220 1.03628692 0.53037136 0.80347656
#> [487] 1.01301889 1.17535629 0.49769518 0.44592241 1.01065544 0.79057353
#> [493] 0.93561445 0.53261400 0.67697358 0.66128056 1.09022740 1.40795352
#> [499] 1.27699758 1.95783607
#> 
#> $dof
#>   [1] 5.426217 5.792707 5.810283 5.742625 5.823476 5.849047 5.917088 5.901018
#>   [9] 5.931500 5.940954 5.907728 5.940566 5.898356 5.891170 5.909418 5.941409
#>  [17] 5.941068 5.884408 5.736212 5.634190 5.851476 5.907256 5.876872 5.925558
#>  [25] 5.919933 5.913751 5.885254 5.770422 5.736937 5.417056 5.338886 4.812448
#>  [33] 5.549862 5.650839 5.890263 5.847942 5.787544 5.321936 5.646815 5.798998
#>  [41] 5.889630 5.834522 5.895364 5.811018 5.713301 5.678022 5.592741 5.655655
#>  [49] 5.408829 5.179066 5.065993 5.228364 5.855938 5.888035 5.912493 5.877046
#>  [57] 5.868925 5.813653 5.805183 5.831879 5.850862 5.893627 5.825571 5.859709
#>  [65] 5.848717 5.792714 5.851813 5.786930 5.699207 5.435722 5.451219 5.662335
#>  [73] 5.851452 5.829227 5.887956 5.917481 5.934205 5.854873 5.768936 5.708725
#>  [81] 5.811587 5.882411 5.831938 5.725823 5.882610 5.894487 5.907705 5.692604
#>  [89] 5.498876 5.515012 5.799578 5.845859 5.810291 5.559532 5.713737 5.638426
#>  [97] 5.689394 5.609987 5.833916 5.742626 5.799843 5.885737 5.882593 5.895877
#> [105] 5.482335 5.168721 5.826483 5.851979 5.882385 5.858543 5.821296 5.771824
#> [113] 5.686321 4.798840 5.781345 5.803881 5.830714 5.810797 5.844710 5.783903
#> [121] 5.808264 5.762544 5.917590 5.940870 5.933818 5.848678 5.432510 5.012943
#> [129] 5.287588 5.535984 5.727381 5.711563 5.795180 5.795806 5.787243 5.777882
#> [137] 5.630690 5.852717 5.823334 5.773755 5.724204 5.786795 5.804636 5.863819
#> [145] 5.908642 5.928947 5.881711 5.892278 5.894438 5.908209 5.927392 5.949397
#> [153] 5.942684 5.910969 5.882067 5.882829 5.912415 5.925649 5.938705 5.931570
#> [161] 5.919351 5.887440 5.905334 5.903087 5.866547 4.799700 5.070178 5.745193
#> [169] 5.902406 5.899629 5.864745 5.670201 5.667897 5.821437 5.802577 5.822929
#> [177] 5.746627 5.742034 5.684562 5.637584 5.571944 5.730169 5.804697 5.809520
#> [185] 5.714779 5.846063 5.766504 5.756665 5.825889 5.827443 5.875183 5.825514
#> [193] 5.771199 4.858521 5.638207 5.869087 5.885997 5.859972 5.681809 5.848876
#> [201] 5.798393 5.891541 5.822385 5.735341 5.674749 5.539348 5.865674 5.815134
#> [209] 5.802073 5.616972 5.043148 5.443512 5.803731 5.807122 5.785062 5.547335
#> [217] 5.407848 5.694765 5.690738 5.631342 5.797121 5.910512 5.918798 5.911089
#> [225] 5.901666 5.934412 5.901111 5.902027 5.814221 5.800835 5.837378 5.863983
#> [233] 5.893147 5.816508 5.704812 5.528912 5.833323 5.932812 5.943188 5.959923
#> [241] 5.917352 5.913427 5.748944 5.712203 5.785113 5.646601 5.787930 5.830948
#> [249] 5.852103 5.787810 5.789650 5.793794 5.884977 5.898818 5.901327 5.837359
#> [257] 5.869583 5.878527 5.898808 5.853795 5.799490 5.809058 5.721411 5.786375
#> [265] 5.870291 5.882639 5.824619 5.857994 5.853011 5.880683 5.764133 5.876617
#> [273] 5.824420 5.840190 5.722339 5.879210 5.747852 5.828548 5.879411 5.856118
#> [281] 5.846694 5.741193 5.839628 5.763681 5.843065 5.846140 5.893132 5.880671
#> [289] 5.830027 5.844165 5.916406 5.887532 5.912128 5.649099 5.787334 5.738859
#> [297] 5.752876 5.665362 5.808874 5.871070 5.845845 5.900117 5.930164 5.924185
#> [305] 5.875870 5.518170 4.934249 5.542177 5.814967 5.909997 5.919489 5.887047
#> [313] 5.785256 5.373672 5.700597 5.932124 5.936602 5.909674 5.908074 5.903577
#> [321] 5.859496 5.765017 5.730716 5.722350 5.683125 5.517499 5.486369 5.380622
#> [329] 5.395626 5.503527 5.746484 5.833328 5.909903 5.908507 5.907415 5.818620
#> [337] 5.731395 5.801394 5.852500 5.855831 5.873065 5.785521 5.822153 5.816671
#> [345] 5.768502 5.734709 5.735973 5.828773 5.868995 5.886969 5.821555 5.762192
#> [353] 5.428629 5.573310 5.871955 5.919919 5.932330 5.892953 5.824756 5.710666
#> [361] 5.731305 5.898684 5.908590 5.887062 5.885250 5.832317 5.839335 5.710083
#> [369] 5.789821 5.770329 5.769561 5.509308 5.433767 5.860758 5.882841 5.918034
#> [377] 5.855297 5.616728 5.372341 5.480456 5.627857 5.755365 5.759472 5.825522
#> [385] 5.877312 5.748906 5.831140 5.832486 5.820353 5.819443 5.917329 5.904092
#> [393] 5.916290 5.633350 5.567791 5.551935 5.814507 5.909840 5.915646 5.919206
#> [401] 5.777492 5.661726 5.729468 5.853183 5.915002 5.827408 5.747234 5.723215
#> [409] 5.609527 5.600086 5.632524 5.637244 5.768384 5.839303 5.863943 5.776252
#> [417] 5.715419 5.578341 5.721422 5.774356 5.860954 5.906302 5.913336 5.886848
#> [425] 5.849166 5.875481 5.876775 5.677468 5.416147 5.700207 5.830303 5.894362
#> [433] 5.890426 5.862234 5.694924 5.532997 5.494610 5.525814 5.626256 5.771156
#> [441] 5.857631 5.910281 5.794170 5.729306 5.761817 5.728280 5.844098 5.724470
#> [449] 5.673261 5.418362 5.798872 5.809091 5.709418 5.665059 5.589049 5.636212
#> [457] 5.821402 5.815778 5.818377 5.605373 5.391976 5.521322 5.551260 5.743163
#> [465] 5.836556 5.877506 5.813479 5.887579 5.843644 5.835771 5.650237 5.834412
#> [473] 5.892027 5.937214 5.904338 5.883786 5.814915 5.552196 5.528412 5.693524
#> [481] 5.795897 5.851345 5.814323 5.834380 5.693045 5.789021 5.830933 5.852949
#> [489] 5.675292 5.642385 5.830138 5.785010 5.817370 5.693997 5.751331 5.746466
#> [497] 5.842222 5.900724 5.864429 5.936179
#> 
#> attr(,".internal.selfref")
#> <pointer: 0x5636ce914340>
#> attr(,"class")
#> [1] "spec" "list"
#> 
#> attr(,"class")
#> [1] "list"      "spec_list"