Transform a spec_df Object into a spec Object or List of spec Objects

Transform a spec_df Object into a spec Object or List of spec Objects

Usage

DF2Spec(spec_df)

Arguments

spec_df

A spec_df object

Value

A spec object or a list of spec objects

See also

Author

Andrew Dolman <andrew.dolman@awi.de>

Examples

library(PaleoSpec)
ts1 <- ts(SimPLS(1000, beta = 1))
sp1 <- SpecMTM(ts1)
ts2 <- ts(rnorm(1000))
sp2 <- SpecMTM(ts2)
sp_lst <- list(sp1 = sp1, sp2 = sp2)
sp_df <- Spec2DF(sp_lst)
sp_df
#> # A tibble: 1,000 × 4
#>    spec_id  freq  spec   dof
#>    <chr>   <dbl> <dbl> <dbl>
#>  1 sp1     0.001 13.9   5.99
#>  2 sp1     0.002 20.3   5.99
#>  3 sp1     0.003 10.2   5.99
#>  4 sp1     0.004 11.5   5.99
#>  5 sp1     0.005  8.47  5.97
#>  6 sp1     0.006  2.67  5.95
#>  7 sp1     0.007  2.56  5.94
#>  8 sp1     0.008  6.25  5.96
#>  9 sp1     0.009  9.67  5.99
#> 10 sp1     0.01   8.85  5.98
#> # … with 990 more rows
DF2Spec(sp_df)
#> $sp1
#> $spec_id
#>   [1] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [13] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [25] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [37] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [49] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [61] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [73] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [85] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#>  [97] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [109] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [121] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [133] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [145] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [157] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [169] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [181] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [193] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [205] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [217] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [229] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [241] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [253] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [265] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [277] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [289] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [301] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [313] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [325] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [337] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [349] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [361] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [373] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [385] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [397] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [409] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [421] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [433] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [445] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [457] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [469] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [481] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> [493] "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1" "sp1"
#> 
#> $freq
#>   [1] 0.001 0.002 0.003 0.004 0.005 0.006 0.007 0.008 0.009 0.010 0.011 0.012
#>  [13] 0.013 0.014 0.015 0.016 0.017 0.018 0.019 0.020 0.021 0.022 0.023 0.024
#>  [25] 0.025 0.026 0.027 0.028 0.029 0.030 0.031 0.032 0.033 0.034 0.035 0.036
#>  [37] 0.037 0.038 0.039 0.040 0.041 0.042 0.043 0.044 0.045 0.046 0.047 0.048
#>  [49] 0.049 0.050 0.051 0.052 0.053 0.054 0.055 0.056 0.057 0.058 0.059 0.060
#>  [61] 0.061 0.062 0.063 0.064 0.065 0.066 0.067 0.068 0.069 0.070 0.071 0.072
#>  [73] 0.073 0.074 0.075 0.076 0.077 0.078 0.079 0.080 0.081 0.082 0.083 0.084
#>  [85] 0.085 0.086 0.087 0.088 0.089 0.090 0.091 0.092 0.093 0.094 0.095 0.096
#>  [97] 0.097 0.098 0.099 0.100 0.101 0.102 0.103 0.104 0.105 0.106 0.107 0.108
#> [109] 0.109 0.110 0.111 0.112 0.113 0.114 0.115 0.116 0.117 0.118 0.119 0.120
#> [121] 0.121 0.122 0.123 0.124 0.125 0.126 0.127 0.128 0.129 0.130 0.131 0.132
#> [133] 0.133 0.134 0.135 0.136 0.137 0.138 0.139 0.140 0.141 0.142 0.143 0.144
#> [145] 0.145 0.146 0.147 0.148 0.149 0.150 0.151 0.152 0.153 0.154 0.155 0.156
#> [157] 0.157 0.158 0.159 0.160 0.161 0.162 0.163 0.164 0.165 0.166 0.167 0.168
#> [169] 0.169 0.170 0.171 0.172 0.173 0.174 0.175 0.176 0.177 0.178 0.179 0.180
#> [181] 0.181 0.182 0.183 0.184 0.185 0.186 0.187 0.188 0.189 0.190 0.191 0.192
#> [193] 0.193 0.194 0.195 0.196 0.197 0.198 0.199 0.200 0.201 0.202 0.203 0.204
#> [205] 0.205 0.206 0.207 0.208 0.209 0.210 0.211 0.212 0.213 0.214 0.215 0.216
#> [217] 0.217 0.218 0.219 0.220 0.221 0.222 0.223 0.224 0.225 0.226 0.227 0.228
#> [229] 0.229 0.230 0.231 0.232 0.233 0.234 0.235 0.236 0.237 0.238 0.239 0.240
#> [241] 0.241 0.242 0.243 0.244 0.245 0.246 0.247 0.248 0.249 0.250 0.251 0.252
#> [253] 0.253 0.254 0.255 0.256 0.257 0.258 0.259 0.260 0.261 0.262 0.263 0.264
#> [265] 0.265 0.266 0.267 0.268 0.269 0.270 0.271 0.272 0.273 0.274 0.275 0.276
#> [277] 0.277 0.278 0.279 0.280 0.281 0.282 0.283 0.284 0.285 0.286 0.287 0.288
#> [289] 0.289 0.290 0.291 0.292 0.293 0.294 0.295 0.296 0.297 0.298 0.299 0.300
#> [301] 0.301 0.302 0.303 0.304 0.305 0.306 0.307 0.308 0.309 0.310 0.311 0.312
#> [313] 0.313 0.314 0.315 0.316 0.317 0.318 0.319 0.320 0.321 0.322 0.323 0.324
#> [325] 0.325 0.326 0.327 0.328 0.329 0.330 0.331 0.332 0.333 0.334 0.335 0.336
#> [337] 0.337 0.338 0.339 0.340 0.341 0.342 0.343 0.344 0.345 0.346 0.347 0.348
#> [349] 0.349 0.350 0.351 0.352 0.353 0.354 0.355 0.356 0.357 0.358 0.359 0.360
#> [361] 0.361 0.362 0.363 0.364 0.365 0.366 0.367 0.368 0.369 0.370 0.371 0.372
#> [373] 0.373 0.374 0.375 0.376 0.377 0.378 0.379 0.380 0.381 0.382 0.383 0.384
#> [385] 0.385 0.386 0.387 0.388 0.389 0.390 0.391 0.392 0.393 0.394 0.395 0.396
#> [397] 0.397 0.398 0.399 0.400 0.401 0.402 0.403 0.404 0.405 0.406 0.407 0.408
#> [409] 0.409 0.410 0.411 0.412 0.413 0.414 0.415 0.416 0.417 0.418 0.419 0.420
#> [421] 0.421 0.422 0.423 0.424 0.425 0.426 0.427 0.428 0.429 0.430 0.431 0.432
#> [433] 0.433 0.434 0.435 0.436 0.437 0.438 0.439 0.440 0.441 0.442 0.443 0.444
#> [445] 0.445 0.446 0.447 0.448 0.449 0.450 0.451 0.452 0.453 0.454 0.455 0.456
#> [457] 0.457 0.458 0.459 0.460 0.461 0.462 0.463 0.464 0.465 0.466 0.467 0.468
#> [469] 0.469 0.470 0.471 0.472 0.473 0.474 0.475 0.476 0.477 0.478 0.479 0.480
#> [481] 0.481 0.482 0.483 0.484 0.485 0.486 0.487 0.488 0.489 0.490 0.491 0.492
#> [493] 0.493 0.494 0.495 0.496 0.497 0.498 0.499 0.500
#> 
#> $spec
#>   [1] 13.85425014 20.33465745 10.15544976 11.49185570  8.47000178  2.67441656
#>   [7]  2.55774619  6.24857226  9.67414136  8.85486277  5.21616168  8.37605567
#>  [13]  8.19566371  6.33220791  5.33266032  1.95513842  4.54140316  5.12616042
#>  [19]  4.54933098  1.58132974  2.87638418  2.24123145  2.33127004  2.16303134
#>  [25]  3.10310326  2.97249483  2.34514966  2.11175289  2.02656620  2.42901968
#>  [31]  1.29576027  2.04763141  0.95250707  0.88601174  0.16788281  0.20633236
#>  [37]  0.19937018  0.95636510  2.44839508  2.91248718  3.07875242  0.49386276
#>  [43]  1.25868715  2.51322833  2.79761904  2.63231174  0.67460677  0.44217767
#>  [49]  1.11315302  1.00474642  1.14242281  0.26615195  0.78151574  1.24325244
#>  [55]  1.39451530  1.42158579  0.99944634  1.55066720  1.09389752  1.05936704
#>  [61]  0.69772628  0.80562082  0.84060495  0.77813976  0.34159017  0.74995486
#>  [67]  1.00376105  0.85020482  0.67547565  0.98577814  0.92081984  1.63084786
#>  [73]  0.96250420  0.97238358  0.99829943  0.45241206  0.55552299  0.27606438
#>  [79]  0.30098615  0.25057213  0.30523712  0.19880045  0.59596113  1.13727250
#>  [85]  1.09125579  0.71984109  0.49156912  1.14502022  1.24644768  1.44588340
#>  [91]  0.63312691  0.78200103  0.84704575  0.50805841  0.52248631  1.07756399
#>  [97]  1.53303619  1.14044800  0.84152012  1.22933797  0.84100887  0.66595712
#> [103]  0.67212515  1.47270070  1.34673155  1.54693836  1.34338699  0.67797744
#> [109]  0.09633847  0.51702114  0.43169139  0.46897683  0.64044911  0.56062637
#> [115]  0.55605371  0.62797056  0.54955368  0.38142902  0.40025260  0.37877120
#> [121]  0.27241656  0.42337114  0.51443199  0.53832506  0.68449395  0.80552526
#> [127]  0.81834549  0.35157146  0.21623957  0.15903925  0.17168086  0.39736854
#> [133]  0.56771442  0.46217087  0.63526122  0.44933369  0.68297728  0.64513629
#> [139]  0.82371380  0.69100151  0.28166558  0.13872090  0.41439785  0.50000670
#> [145]  0.47030008  0.29666803  0.37869900  0.40408732  0.34032532  0.16726867
#> [151]  0.34602910  0.29386463  0.54478472  0.41237324  0.43299272  0.20709556
#> [157]  0.24697477  0.35837689  0.41515203  0.65326337  0.65394470  0.86616187
#> [163]  0.27952559  0.25426508  0.54300142  0.60311060  0.64904434  0.40472027
#> [169]  0.23840316  0.25321290  0.38115325  0.43931183  0.43162825  0.39412432
#> [175]  0.09188257  0.05476684  0.27121777  0.34635421  0.38388393  0.40301243
#> [181]  0.35224073  0.50100895  0.54077307  0.27367137  0.23621294  0.19173164
#> [187]  0.36245239  0.56879171  0.66736133  0.61804062  0.70611132  0.40410826
#> [193]  0.33686558  0.25753514  0.33881377  0.54042135  0.52085802  0.55538085
#> [199]  0.62026496  0.30809847  0.24713206  0.33944381  0.30513965  0.27467147
#> [205]  0.09037302  0.56883402  0.58437956  0.74287245  0.26836537  0.38759586
#> [211]  0.28217155  0.06537379  0.06368667  0.31973552  0.72783366  0.72123328
#> [217]  0.70969969  0.58222375  0.58688217  0.54511048  0.35293432  0.51209218
#> [223]  0.54593870  0.49576901  0.19655019  0.15641279  0.25920735  0.24529293
#> [229]  0.26593587  0.29567326  0.61969163  0.56699886  0.45250468  0.38952233
#> [235]  0.23933038  0.27189384  0.27995207  0.15872985  0.16744693  0.12167872
#> [241]  0.03056034  0.10695105  0.16394289  0.16732931  0.16384350  0.19160368
#> [247]  0.37530746  0.49840408  0.55147782  0.32248484  0.32309871  0.36826746
#> [253]  0.25320822  0.32111047  0.24879996  0.29457347  0.30508530  0.23781656
#> [259]  0.20539417  0.15884301  0.11677494  0.25829270  0.29685675  0.29239286
#> [265]  0.31434865  0.26255207  0.34525983  0.23494827  0.23537116  0.14770926
#> [271]  0.15365609  0.19378244  0.26634611  0.35083335  0.31624863  0.26451554
#> [277]  0.24989058  0.29885022  0.39203066  0.31580108  0.30065141  0.10757156
#> [283]  0.12129974  0.23505949  0.16351013  0.14640703  0.19666012  0.35746354
#> [289]  0.35310517  0.32642714  0.23079553  0.29015826  0.50720907  0.56310780
#> [295]  0.36782803  0.28385103  0.50858769  0.63013272  0.60922233  0.56244645
#> [301]  0.33428990  0.38933485  0.27556953  0.31341347  0.29243288  0.29021952
#> [307]  0.18639088  0.02023006  0.01635004  0.07476393  0.09195694  0.13595126
#> [313]  0.21631629  0.41415209  0.36360962  0.29411982  0.29971876  0.26001434
#> [319]  0.31462160  0.20975185  0.13816856  0.06844718  0.08888375  0.06755785
#> [325]  0.24725689  0.28375272  0.26953437  0.19764616  0.12691139  0.19344076
#> [331]  0.17937381  0.22007446  0.36200133  0.40830825  0.35328881  0.35911355
#> [337]  0.25212668  0.27726435  0.23591925  0.17190168  0.08400053  0.16144814
#> [343]  0.18629586  0.28210523  0.31919520  0.47385703  0.31163916  0.27755448
#> [349]  0.47209014  0.49548798  0.58256675  0.40916581  0.27829789  0.16964286
#> [355]  0.12011492  0.11767763  0.12043044  0.04173559  0.18138758  0.25950403
#> [361]  0.25431766  0.22789626  0.31173967  0.28966715  0.23343929  0.12740581
#> [367]  0.12502364  0.07872688  0.08869281  0.11114666  0.09872391  0.12233420
#> [373]  0.19958912  0.16194857  0.20715028  0.14746226  0.13988834  0.06610358
#> [379]  0.16323325  0.39599671  0.40388512  0.18674275  0.08855937  0.05746995
#> [385]  0.09859840  0.11553245  0.23953653  0.19188110  0.15268032  0.06267833
#> [391]  0.05179576  0.11532352  0.10477598  0.10947054  0.06187627  0.12892707
#> [397]  0.39556079  0.44126216  0.37111191  0.11004643  0.01717939  0.03784966
#> [403]  0.04869479  0.06706699  0.04107956  0.15995144  0.26171757  0.33027855
#> [409]  0.20933465  0.15582171  0.18519013  0.22727741  0.22791958  0.25233790
#> [415]  0.19771428  0.29309488  0.26541025  0.24140697  0.14516530  0.10117779
#> [421]  0.09143791  0.09092332  0.14548867  0.18053127  0.10470898  0.04954758
#> [427]  0.01291411  0.17739320  0.19505767  0.18875643  0.18674290  0.26410361
#> [433]  0.45402698  0.38989966  0.19302335  0.05460452  0.04879279  0.05865525
#> [439]  0.03861414  0.05160745  0.10733178  0.14161462  0.16363914  0.12021959
#> [445]  0.11313680  0.08566577  0.08173302  0.04015994  0.03055825  0.04541050
#> [451]  0.04783533  0.06869726  0.09814028  0.11006667  0.09726505  0.10791235
#> [457]  0.19885700  0.22630845  0.29062272  0.39277524  0.38028856  0.35896725
#> [463]  0.23678730  0.03458811  0.03259841  0.03964205  0.02711852  0.03468121
#> [469]  0.03639039  0.04027598  0.14444171  0.21505048  0.27444470  0.13733073
#> [475]  0.13343106  0.30349456  0.35702141  0.49758676  0.43447635  0.26814331
#> [481]  0.18636699  0.07021332  0.15716445  0.15181892  0.17635986  0.11474087
#> [487]  0.20893987  0.24000352  0.21504347  0.17218676  0.18559314  0.13977364
#> [493]  0.06803962  0.11361629  0.13535955  0.11194844  0.02410807  0.04915807
#> [499]  0.07147593  0.09237386
#> 
#> $dof
#>   [1] 5.992880 5.992784 5.990019 5.990041 5.973092 5.946606 5.943573 5.960650
#>   [9] 5.988313 5.981975 5.975714 5.956084 5.984447 5.984217 5.965681 5.949277
#>  [17] 5.975538 5.962774 5.976060 5.895072 5.907125 5.922787 5.940128 5.940864
#>  [25] 5.966945 5.966455 5.916551 5.896299 5.946381 5.942146 5.923707 5.930640
#>  [33] 5.854750 5.879287 5.328085 5.431465 5.414345 5.855019 5.957889 5.952063
#>  [41] 5.934737 5.732135 5.887382 5.953701 5.950480 5.948745 5.798069 5.702558
#>  [49] 5.903035 5.898804 5.877819 5.539214 5.824703 5.885895 5.889107 5.925941
#>  [57] 5.894313 5.908500 5.902377 5.900647 5.804861 5.876742 5.876865 5.824279
#>  [65] 5.627192 5.815414 5.860885 5.838200 5.798617 5.859536 5.850042 5.913142
#>  [73] 5.899446 5.893981 5.861163 5.709701 5.759279 5.549465 5.579248 5.513919
#>  [81] 5.581447 5.410352 5.774452 5.876744 5.870998 5.810008 5.730349 5.894772
#>  [89] 5.905775 5.907547 5.785255 5.823815 5.837270 5.738434 5.745385 5.870203
#>  [97] 5.898361 5.892302 5.836732 5.885369 5.878444 5.794903 5.797296 5.903911
#> [105] 5.916866 5.935423 5.893436 5.800038 5.005526 5.743100 5.693851 5.717203
#> [113] 5.787945 5.760429 5.759603 5.784986 5.755129 5.661303 5.676430 5.653793
#> [121] 5.547163 5.692362 5.740232 5.751569 5.801645 5.829813 5.831403 5.637319
#> [129] 5.451511 5.303038 5.339380 5.671787 5.763216 5.712382 5.787280 5.707917
#> [137] 5.800545 5.789018 5.833378 5.801630 5.555974 5.227956 5.681976 5.734489
#> [145] 5.719507 5.578453 5.657217 5.679303 5.623763 5.326582 5.629926 5.575783
#> [153] 5.753877 5.681645 5.698424 5.433835 5.508905 5.643376 5.685653 5.792782
#> [161] 5.793229 5.840872 5.557384 5.521651 5.753334 5.774869 5.791384 5.677740
#> [169] 5.494833 5.518502 5.660535 5.700926 5.697632 5.671589 4.975400 4.615687
#> [177] 5.545526 5.627561 5.661863 5.678380 5.635139 5.735540 5.752958 5.545885
#> [185] 5.487817 5.392883 5.646020 5.764479 5.796932 5.781988 5.807186 5.677440
#> [193] 5.621616 5.526664 5.625154 5.751097 5.742016 5.757778 5.782827 5.593235
#> [201] 5.503409 5.625005 5.589744 5.551158 4.963750 5.764771 5.769536 5.816336
#> [209] 5.542040 5.667208 5.560748 4.740905 4.717326 5.603078 5.811784 5.811239
#> [217] 5.806926 5.768149 5.771445 5.754356 5.637706 5.737887 5.755462 5.733216
#> [225] 5.405550 5.292342 5.528542 5.505260 5.537097 5.578539 5.780920 5.763488
#> [233] 5.707540 5.668326 5.493586 5.546786 5.558516 5.302866 5.330588 5.151175
#> [241] 4.198137 5.071998 5.319994 5.324501 5.316726 5.396894 5.657575 5.731703
#> [249] 5.756605 5.606318 5.609500 5.647747 5.516192 5.606757 5.512301 5.573544
#> [257] 5.589421 5.489272 5.430375 5.302452 5.127134 5.523349 5.580484 5.574921
#> [265] 5.599222 5.533829 5.626369 5.487673 5.481184 5.257120 5.283983 5.398826
#> [273] 5.537248 5.636623 5.602024 5.535732 5.514552 5.572041 5.666849 5.596811
#> [281] 5.584506 5.077513 5.140376 5.485775 5.314758 5.257617 5.404619 5.642563
#> [289] 5.635464 5.612721 5.480388 5.571674 5.738607 5.762399 5.651233 5.561269
#> [297] 5.739253 5.784725 5.779199 5.762280 5.619897 5.668523 5.552344 5.599475
#> [305] 5.569043 5.572163 5.381317 3.925268 3.787814 4.836041 4.977381 5.217263
#> [313] 5.453212 5.685155 5.647751 5.573704 5.582172 5.527348 5.597241 5.437267
#> [321] 5.224539 4.773917 4.953244 4.760542 5.508647 5.562746 5.542008 5.402150
#> [329] 5.167897 5.402180 5.365717 5.456828 5.644760 5.681263 5.634707 5.642131
#> [337] 5.517549 5.549820 5.490871 5.341380 4.905398 5.311541 5.382994 5.557736
#> [345] 5.604120 5.721679 5.591461 5.555147 5.720656 5.732298 5.767767 5.679922
#> [353] 5.555924 5.335059 5.132482 5.131185 5.142855 4.421901 5.369524 5.527495
#> [361] 5.519183 5.476006 5.596502 5.568088 5.486015 5.179644 5.167440 4.864948
#> [369] 4.950816 5.096928 5.020075 5.151798 5.415451 5.313907 5.433052 5.263899
#> [377] 5.232855 4.740003 5.315940 5.673233 5.679010 5.381329 4.931958 4.647306
#> [385] 5.017899 5.121115 5.494901 5.397541 5.281616 4.708722 4.577080 5.119575
#> [393] 5.059651 5.088780 4.685358 5.186351 5.672781 5.702497 5.653798 5.083850
#> [401] 3.819488 4.346147 4.531129 4.759794 4.408105 5.306706 5.531748 5.616594
#> [409] 5.438570 5.292755 5.380007 5.472824 5.475644 5.518358 5.412285 5.575379
#> [417] 5.537705 5.499118 5.250812 5.037350 4.970832 4.967896 5.254478 5.367216
#> [425] 5.057219 4.542320 3.631698 5.359304 5.400028 5.390214 5.384635 5.531843
#> [433] 5.709597 5.668427 5.399103 4.613947 4.534533 4.665463 4.364453 4.573007
#> [441] 5.076174 5.240625 5.314608 5.145111 5.106799 4.923798 4.893697 4.393932
#> [449] 4.202264 4.481531 4.520450 4.776616 5.014764 5.089620 5.012147 5.078602
#> [457] 5.413590 5.473282 5.571559 5.670970 5.657404 5.641007 5.491503 4.286981
#> [465] 4.225516 4.383167 4.120075 4.289757 4.319065 4.380793 5.248188 5.450342
#> [473] 5.549409 5.223165 5.205660 5.587703 5.637804 5.733192 5.699373 5.538601
#> [481] 5.383652 4.790277 5.296582 5.278322 5.357257 5.113120 5.436540 5.493755
#> [489] 5.450338 5.342511 5.381913 5.232140 4.768830 5.110356 5.213929 5.101584
#> [497] 4.020213 4.538372 4.805782 4.979258
#> 
#> attr(,".internal.selfref")
#> <pointer: 0x56369671a340>
#> attr(,"class")
#> [1] "spec" "list"
#> 
#> $sp2
#> $spec_id
#>   [1] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [13] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [25] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [37] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [49] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [61] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [73] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [85] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#>  [97] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [109] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [121] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [133] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [145] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [157] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [169] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [181] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [193] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [205] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [217] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [229] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [241] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [253] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [265] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [277] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [289] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [301] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [313] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [325] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [337] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [349] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [361] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [373] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [385] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [397] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [409] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [421] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [433] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [445] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [457] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [469] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [481] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> [493] "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2" "sp2"
#> 
#> $freq
#>   [1] 0.001 0.002 0.003 0.004 0.005 0.006 0.007 0.008 0.009 0.010 0.011 0.012
#>  [13] 0.013 0.014 0.015 0.016 0.017 0.018 0.019 0.020 0.021 0.022 0.023 0.024
#>  [25] 0.025 0.026 0.027 0.028 0.029 0.030 0.031 0.032 0.033 0.034 0.035 0.036
#>  [37] 0.037 0.038 0.039 0.040 0.041 0.042 0.043 0.044 0.045 0.046 0.047 0.048
#>  [49] 0.049 0.050 0.051 0.052 0.053 0.054 0.055 0.056 0.057 0.058 0.059 0.060
#>  [61] 0.061 0.062 0.063 0.064 0.065 0.066 0.067 0.068 0.069 0.070 0.071 0.072
#>  [73] 0.073 0.074 0.075 0.076 0.077 0.078 0.079 0.080 0.081 0.082 0.083 0.084
#>  [85] 0.085 0.086 0.087 0.088 0.089 0.090 0.091 0.092 0.093 0.094 0.095 0.096
#>  [97] 0.097 0.098 0.099 0.100 0.101 0.102 0.103 0.104 0.105 0.106 0.107 0.108
#> [109] 0.109 0.110 0.111 0.112 0.113 0.114 0.115 0.116 0.117 0.118 0.119 0.120
#> [121] 0.121 0.122 0.123 0.124 0.125 0.126 0.127 0.128 0.129 0.130 0.131 0.132
#> [133] 0.133 0.134 0.135 0.136 0.137 0.138 0.139 0.140 0.141 0.142 0.143 0.144
#> [145] 0.145 0.146 0.147 0.148 0.149 0.150 0.151 0.152 0.153 0.154 0.155 0.156
#> [157] 0.157 0.158 0.159 0.160 0.161 0.162 0.163 0.164 0.165 0.166 0.167 0.168
#> [169] 0.169 0.170 0.171 0.172 0.173 0.174 0.175 0.176 0.177 0.178 0.179 0.180
#> [181] 0.181 0.182 0.183 0.184 0.185 0.186 0.187 0.188 0.189 0.190 0.191 0.192
#> [193] 0.193 0.194 0.195 0.196 0.197 0.198 0.199 0.200 0.201 0.202 0.203 0.204
#> [205] 0.205 0.206 0.207 0.208 0.209 0.210 0.211 0.212 0.213 0.214 0.215 0.216
#> [217] 0.217 0.218 0.219 0.220 0.221 0.222 0.223 0.224 0.225 0.226 0.227 0.228
#> [229] 0.229 0.230 0.231 0.232 0.233 0.234 0.235 0.236 0.237 0.238 0.239 0.240
#> [241] 0.241 0.242 0.243 0.244 0.245 0.246 0.247 0.248 0.249 0.250 0.251 0.252
#> [253] 0.253 0.254 0.255 0.256 0.257 0.258 0.259 0.260 0.261 0.262 0.263 0.264
#> [265] 0.265 0.266 0.267 0.268 0.269 0.270 0.271 0.272 0.273 0.274 0.275 0.276
#> [277] 0.277 0.278 0.279 0.280 0.281 0.282 0.283 0.284 0.285 0.286 0.287 0.288
#> [289] 0.289 0.290 0.291 0.292 0.293 0.294 0.295 0.296 0.297 0.298 0.299 0.300
#> [301] 0.301 0.302 0.303 0.304 0.305 0.306 0.307 0.308 0.309 0.310 0.311 0.312
#> [313] 0.313 0.314 0.315 0.316 0.317 0.318 0.319 0.320 0.321 0.322 0.323 0.324
#> [325] 0.325 0.326 0.327 0.328 0.329 0.330 0.331 0.332 0.333 0.334 0.335 0.336
#> [337] 0.337 0.338 0.339 0.340 0.341 0.342 0.343 0.344 0.345 0.346 0.347 0.348
#> [349] 0.349 0.350 0.351 0.352 0.353 0.354 0.355 0.356 0.357 0.358 0.359 0.360
#> [361] 0.361 0.362 0.363 0.364 0.365 0.366 0.367 0.368 0.369 0.370 0.371 0.372
#> [373] 0.373 0.374 0.375 0.376 0.377 0.378 0.379 0.380 0.381 0.382 0.383 0.384
#> [385] 0.385 0.386 0.387 0.388 0.389 0.390 0.391 0.392 0.393 0.394 0.395 0.396
#> [397] 0.397 0.398 0.399 0.400 0.401 0.402 0.403 0.404 0.405 0.406 0.407 0.408
#> [409] 0.409 0.410 0.411 0.412 0.413 0.414 0.415 0.416 0.417 0.418 0.419 0.420
#> [421] 0.421 0.422 0.423 0.424 0.425 0.426 0.427 0.428 0.429 0.430 0.431 0.432
#> [433] 0.433 0.434 0.435 0.436 0.437 0.438 0.439 0.440 0.441 0.442 0.443 0.444
#> [445] 0.445 0.446 0.447 0.448 0.449 0.450 0.451 0.452 0.453 0.454 0.455 0.456
#> [457] 0.457 0.458 0.459 0.460 0.461 0.462 0.463 0.464 0.465 0.466 0.467 0.468
#> [469] 0.469 0.470 0.471 0.472 0.473 0.474 0.475 0.476 0.477 0.478 0.479 0.480
#> [481] 0.481 0.482 0.483 0.484 0.485 0.486 0.487 0.488 0.489 0.490 0.491 0.492
#> [493] 0.493 0.494 0.495 0.496 0.497 0.498 0.499 0.500
#> 
#> $spec
#>   [1] 0.69374081 1.14195811 1.68650331 0.87227647 1.30901481 1.53771562
#>   [7] 1.58227352 1.56843214 1.19486665 0.75532072 1.26705532 1.01355378
#>  [13] 0.76722693 0.44938820 0.70573854 0.63177717 1.88587770 1.88180836
#>  [19] 1.41211031 0.95193793 0.28352918 0.69602411 0.91005570 1.39333953
#>  [25] 1.62974483 0.64335263 0.97818912 0.78461893 0.80183854 0.39661971
#>  [31] 0.39948618 0.19007606 0.18304596 0.22319592 0.12934407 0.08356295
#>  [37] 0.61536303 0.67283359 0.92176833 0.70302369 0.84071056 0.53844300
#>  [43] 0.60612736 0.62262761 1.87278052 2.05185098 1.84241764 0.10824815
#>  [49] 0.17834249 0.23767307 0.28984386 0.41083219 0.51668513 1.16440171
#>  [55] 2.73671039 2.75698426 2.78716425 1.30632548 0.67930512 0.56008697
#>  [61] 0.43658623 1.19730431 1.28806374 1.82629660 2.90622047 1.78935976
#>  [67] 1.70880874 0.49088796 0.47040254 0.39337215 0.29259853 1.40432769
#>  [73] 1.22980595 2.22271846 2.19494324 1.42467632 1.41202089 0.43294716
#>  [79] 0.52529449 0.72418393 0.69950459 0.38535561 0.91669168 0.92311986
#>  [85] 1.17986173 0.43485601 0.93039746 1.35023890 1.14752046 1.26775265
#>  [91] 1.08651782 1.16781610 1.03401462 1.04658406 1.05421393 0.47090758
#>  [97] 1.42076829 1.35798035 2.19854325 1.04546079 1.14915485 0.77243723
#> [103] 1.21794963 1.33343067 1.08939337 0.15617344 0.23482205 0.68636868
#> [109] 0.55102047 0.73707054 0.70528930 0.53767456 0.63978001 1.12952069
#> [115] 1.23857987 0.58235776 0.46681062 0.70551792 1.15689823 1.47199294
#> [121] 0.98595450 1.31886218 0.72384127 0.86561554 0.65259925 0.96260084
#> [127] 0.92337559 0.84039143 0.70556501 0.07941142 0.15257760 0.16075691
#> [133] 0.16038822 0.29273857 0.73144612 1.35060796 1.50542424 1.57263314
#> [139] 0.80861548 0.92170745 0.98231194 0.57366046 0.43678077 0.43996326
#> [145] 0.62145265 0.71790562 0.61808492 0.47391432 0.10908329 0.17645049
#> [151] 0.39335385 1.37139945 1.34292299 1.05983305 1.60191176 1.47756315
#> [157] 1.62301601 1.24533490 1.05524775 2.03284175 2.12893981 1.15717687
#> [163] 0.86839515 0.21612592 0.32981775 0.69748227 0.71814152 1.43716059
#> [169] 1.09128019 1.63147296 1.77846640 2.33944121 2.15781175 1.89283831
#> [175] 0.37154894 0.18562916 0.44312370 1.19233273 1.12842360 0.53999928
#> [181] 1.03342885 1.26445261 1.73523622 1.13901814 1.62332244 1.44291417
#> [187] 0.55182943 0.49514717 0.63579652 0.42451233 1.38180526 1.34713078
#> [193] 1.44953782 1.33072227 0.64035702 0.47706628 0.43888952 0.71647925
#> [199] 0.65282799 0.65170259 0.50078103 0.22059158 0.48699192 0.54817640
#> [205] 1.03795349 1.50855800 2.40381943 3.04510822 1.92231858 1.63252434
#> [211] 0.63530371 0.63174951 0.62043874 0.60493180 0.67674621 1.71223281
#> [217] 2.15648982 2.58930116 2.61199288 2.10240118 1.33140083 0.50241207
#> [223] 0.53680271 0.53389633 1.24219590 0.88604574 1.84454324 1.48723888
#> [229] 1.25582084 0.72971469 0.17493769 1.19158778 2.16632303 2.20800981
#> [235] 2.56966941 1.45727147 1.07394661 1.31159289 1.47573340 1.61429462
#> [241] 1.57163353 0.85076388 0.43112122 0.20668087 0.10658293 0.43566091
#> [247] 1.28300113 1.50422580 1.21842887 1.03491761 0.91283090 0.22737426
#> [253] 0.40509023 0.42862319 0.55058318 0.59923722 0.66289500 0.69946087
#> [259] 0.81233493 0.73587361 0.86842572 0.48644555 0.44703834 0.35841094
#> [265] 0.40311792 0.59915487 0.53744373 0.54732773 0.09695296 0.21432763
#> [271] 0.20314484 0.57923804 1.13914575 2.07944354 1.91314709 2.13251558
#> [277] 2.52665817 1.10594113 0.76839079 0.75992154 0.45375537 0.55914427
#> [283] 0.55076929 0.42750145 0.67891884 0.61100728 0.54342535 0.35203782
#> [289] 0.61843536 0.66016263 0.85906853 1.19774496 1.74276821 1.43787960
#> [295] 1.23518556 0.44542204 1.67135789 1.71159938 1.86847149 0.97477587
#> [301] 0.51199588 0.84018525 0.84842879 1.38409732 1.24770097 1.45216743
#> [307] 1.48348344 2.26062607 1.78066863 1.46148431 1.31829015 1.36506314
#> [313] 0.37737952 0.65058852 0.89975102 1.00973534 0.61862276 0.49077526
#> [319] 0.62777422 0.71682273 0.90643894 1.20574897 1.00142487 0.95446275
#> [325] 0.48947809 1.50716931 1.54937277 1.57051714 0.72279871 0.34797644
#> [331] 0.27817547 0.56678300 0.62232681 1.05356870 0.98057610 0.90614942
#> [337] 0.39015321 0.61502043 0.53845138 0.50577471 1.13571061 1.29489066
#> [343] 1.56875530 1.04533081 1.14640905 1.17145824 1.78755635 1.66085632
#> [349] 1.45632247 0.88865180 1.04415115 2.18140439 2.26324337 2.90782983
#> [355] 0.93797195 0.71635452 0.57076250 0.29234807 0.99322529 1.29773886
#> [361] 1.83830242 2.10336845 1.47098306 1.42889575 0.78602359 0.53375695
#> [367] 0.77544277 1.03978649 0.98735045 0.50190003 0.43962040 1.25888954
#> [373] 1.24893721 1.14551086 0.14090533 0.29314211 0.75262475 1.35048930
#> [379] 2.31257177 2.52730277 1.67474583 1.34878752 1.23321145 0.85261328
#> [385] 0.66297473 0.50635681 0.85302451 0.70124707 1.07358394 0.84624125
#> [391] 0.99908259 1.51109301 1.51887603 1.22876111 1.42775050 1.88500433
#> [397] 1.90333916 2.68555677 1.44687426 1.40731515 1.15784521 1.05518531
#> [403] 1.10147432 0.91390831 1.23885795 1.20642539 1.32006699 0.40876934
#> [409] 0.37996414 0.48579792 0.72886723 0.64498616 0.82427780 0.66137055
#> [415] 0.37922859 0.35958845 0.24722357 0.31180442 1.31653937 1.89618525
#> [421] 2.02941757 2.20352372 0.96504017 0.75736146 0.87917297 2.04985432
#> [427] 2.40948551 2.83453416 1.73750882 0.81263276 0.92835155 1.24693844
#> [433] 2.19036421 1.39827091 2.46124202 2.11701955 1.44531191 1.45406992
#> [439] 1.82251648 1.90250798 0.71102062 0.33243794 0.53200700 1.38398740
#> [445] 1.39774663 1.28299691 1.01791438 1.17417094 1.17727781 0.86605167
#> [451] 0.39946731 0.51701955 0.65071594 1.09721618 1.61663130 0.77437628
#> [457] 0.71370075 0.75922304 0.70397615 3.01156028 2.31596380 2.84154389
#> [463] 1.18849519 0.79171287 0.24764660 0.23139144 0.36387345 0.18537400
#> [469] 0.32443963 1.11652824 1.46092189 1.83980121 1.58328644 1.33933506
#> [475] 0.64120659 1.07566727 1.09304126 1.09362294 0.78229455 0.65610506
#> [481] 0.96704742 2.09477128 1.54284690 1.44998209 0.59976270 0.20329900
#> [487] 0.53329923 0.86623501 1.85159567 1.97731896 1.94866736 1.79108252
#> [493] 0.73965187 0.70544637 0.63236394 1.08304947 0.88881590 0.76677460
#> [499] 1.20160908 1.61172815
#> 
#> $dof
#>   [1] 5.746028 5.840723 5.901214 5.794581 5.859871 5.889188 5.909992 5.891730
#>   [9] 5.847333 5.764585 5.855706 5.821771 5.768406 5.625478 5.750213 5.723900
#>  [17] 5.905409 5.903852 5.908749 5.808509 5.446868 5.747190 5.801547 5.876789
#>  [25] 5.893824 5.728614 5.809833 5.772559 5.778289 5.583285 5.583928 5.248852
#>  [33] 5.227871 5.334982 5.010340 4.716899 5.715514 5.739529 5.803717 5.749638
#>  [41] 5.787463 5.680575 5.710950 5.719760 5.920888 5.908280 5.899189 4.893051
#>  [49] 5.212472 5.354043 5.459114 5.594599 5.663170 5.842177 5.946272 5.949754
#>  [57] 5.939389 5.857536 5.741486 5.690720 5.616205 5.847041 5.898218 5.920718
#>  [65] 5.935395 5.899641 5.891537 5.651262 5.637870 5.580262 5.460476 5.868827
#>  [73] 5.889926 5.924631 5.902235 5.896335 5.869789 5.611446 5.671164 5.756007
#>  [81] 5.747844 5.572437 5.804018 5.805461 5.843252 5.613586 5.806253 5.864113
#>  [89] 5.841389 5.855768 5.832356 5.891524 5.824140 5.826517 5.828065 5.638733
#>  [97] 5.877755 5.864913 5.915063 5.826921 5.887866 5.770502 5.885583 5.861189
#> [105] 5.833504 5.133031 5.358885 5.744023 5.686995 5.758957 5.749247 5.679090
#> [113] 5.727066 5.838327 5.849919 5.703064 5.638673 5.750406 5.842687 5.874930
#> [121] 5.816315 5.861083 5.755376 5.792981 5.730751 5.812084 5.804969 5.786068
#> [129] 5.749974 4.677509 5.116310 5.151175 5.147792 5.462549 5.757743 5.864195
#> [137] 5.909924 5.892865 5.779401 5.805041 5.816422 5.699052 5.616407 5.614000
#> [145] 5.720009 5.752810 5.718618 5.643420 4.902563 5.206544 5.578808 5.865964
#> [153] 5.863270 5.829037 5.895322 5.901179 5.918047 5.851767 5.879976 5.908043
#> [161] 5.928823 5.891242 5.794065 5.318135 5.509044 5.747828 5.754599 5.870810
#> [169] 5.833675 5.890176 5.927363 5.929744 5.935544 5.900837 5.559498 5.235416
#> [177] 5.617094 5.847085 5.838732 5.681565 5.824727 5.884090 5.892884 5.887777
#> [185] 5.885991 5.872550 5.687996 5.652626 5.723758 5.607195 5.867151 5.889470
#> [193] 5.912927 5.861643 5.725329 5.645121 5.618699 5.751446 5.729815 5.729397
#> [201] 5.658842 5.328803 5.650162 5.685943 5.823526 5.889149 5.928311 5.943238
#> [209] 5.925260 5.885212 5.725091 5.723941 5.719628 5.710419 5.740732 5.890522
#> [217] 5.913054 5.933427 5.894521 5.931513 5.862345 5.660276 5.680074 5.679044
#> [225] 5.852407 5.797713 5.923183 5.898261 5.895458 5.758040 5.201459 5.846972
#> [233] 5.906115 5.942575 5.893587 5.909998 5.831117 5.858928 5.908938 5.918764
#> [241] 5.912628 5.790014 5.609287 5.294529 4.888274 5.613121 5.857324 5.875247
#> [249] 5.893959 5.825200 5.800508 5.344050 5.591120 5.609692 5.686988 5.710683
#> [257] 5.735502 5.747457 5.780924 5.758390 5.793120 5.651370 5.622170 5.546328
#> [265] 5.585857 5.710107 5.678979 5.686140 4.819427 5.313599 5.283340 5.701695
#> [273] 5.840270 5.910027 5.926420 5.915028 5.925817 5.835693 5.769390 5.765574
#> [281] 5.624326 5.690624 5.686851 5.609681 5.741615 5.715662 5.680552 5.534409
#> [289] 5.718054 5.734975 5.792022 5.847571 5.893453 5.906589 5.852178 5.623120
#> [297] 5.888385 5.923260 5.919350 5.814522 5.666243 5.787228 5.789169 5.890753
#> [305] 5.853583 5.909868 5.903499 5.917241 5.894663 5.913803 5.860782 5.865475
#> [313] 5.566087 5.729392 5.799804 5.819233 5.716867 5.650025 5.722598 5.753832
#> [321] 5.799934 5.848637 5.819754 5.811070 5.651742 5.900262 5.885682 5.901207
#> [329] 5.754841 5.534572 5.437312 5.695598 5.719997 5.827632 5.816017 5.801876
#> [337] 5.575899 5.716666 5.680865 5.663087 5.837361 5.858611 5.882300 5.826612
#> [345] 5.841103 5.841975 5.901756 5.909730 5.912412 5.798249 5.826147 5.934806
#> [353] 5.929204 5.888250 5.808365 5.753855 5.697610 5.462694 5.815383 5.902506
#> [361] 5.915808 5.894740 5.899465 5.871057 5.772238 5.678873 5.771274 5.825695
#> [369] 5.816894 5.660380 5.617735 5.854731 5.877635 5.841039 5.070939 5.459955
#> [377] 5.761029 5.905573 5.920410 5.911939 5.908099 5.863867 5.884334 5.788484
#> [385] 5.735920 5.661345 5.790553 5.747582 5.831132 5.789011 5.819329 5.877813
#> [393] 5.912258 5.884268 5.871250 5.925225 5.912589 5.914361 5.903175 5.908915
#> [401] 5.842737 5.827900 5.835205 5.803599 5.852557 5.848319 5.861242 5.589821
#> [409] 5.566306 5.648815 5.757541 5.728793 5.782214 5.735238 5.567848 5.541766
#> [417] 5.385247 5.490549 5.860696 5.924734 5.936026 5.914702 5.813228 5.766242
#> [425] 5.796109 5.932414 5.947196 5.943553 5.880036 5.780435 5.805067 5.853437
#> [433] 5.913160 5.898756 5.913652 5.915138 5.872748 5.873004 5.877726 5.902114
#> [441] 5.750894 5.509190 5.676585 5.903227 5.908434 5.857349 5.822325 5.844012
#> [449] 5.844518 5.793440 5.585708 5.669566 5.731481 5.832350 5.885664 5.769282
#> [457] 5.752588 5.765680 5.748913 5.884052 5.944296 5.939077 5.846396 5.774291
#> [465] 5.384113 5.352922 5.550703 5.233593 5.505083 5.836841 5.874037 5.917085
#> [473] 5.881282 5.863147 5.725686 5.831473 5.883073 5.880548 5.773104 5.733425
#> [481] 5.813736 5.894080 5.914810 5.893114 5.710794 5.283410 5.678550 5.792759
#> [489] 5.920285 5.931921 5.899736 5.895354 5.759773 5.748666 5.722207 5.830795
#> [497] 5.797508 5.767895 5.847801 5.884958
#> 
#> attr(,".internal.selfref")
#> <pointer: 0x56369671a340>
#> attr(,"class")
#> [1] "spec" "list"
#> 
#> attr(,"class")
#> [1] "list"      "spec_list"